feat: move rooms and entrance initialization to Resources.cs
Note: this is broken, since Entrances get modified
